refactor(content_ui): localize menu graphql
