+++ frontend/plugins/content_ui/src/modules/cms/menus/components/MenusColumn.tsx
@@ -19,7 +19,7 @@ import {
 import { useState } from 'react';
 import { useConfirm } from 'erxes-ui/hooks/use-confirm';
 import { useMutation } from '@apollo/client';
-import { CMS_MENU_EDIT, CMS_MENU_REMOVE } from '../../graphql/queries';
+import { CMS_MENU_EDIT, CMS_MENU_REMOVE } from '@/cms/menus/graphql';
 import { getDepthPrefix } from '@/cms/menus/utils/menuUtils';
 import { useIsTranslationMissing } from '../../shared/hooks/useIsTranslationMissing';
 import { MenuItem } from '../types/menuDrawerTypes';
