+++ frontend/plugins/content_ui/src/modules/cms/menus/components/MenusRecordTable.tsx
@@ -32,7 +32,7 @@ import { useAtomValue } from 'jotai';
 import { useMenusColumns } from './MenusColumn';
 import { MenusCommandBar } from './menus-command-bar/MenusCommandBar';
 import { useMenus } from '../hooks/useMenus';
-import { CMS_MENU_EDIT, CMS_MENU_REMOVE } from '../../graphql/queries';
+import { CMS_MENU_EDIT, CMS_MENU_REMOVE } from '@/cms/menus/graphql';
 import { buildFlatTree, getDepthPrefix } from '@/cms/menus/utils/menuUtils';
 import { cmsLanguageAtom } from '@/cms/shared/states/cmsLanguageState';
 import { MenuItem } from '../types/menuDrawerTypes';
