+++ frontend/plugins/content_ui/src/modules/cms/menus/graphql/mutations/index.ts
@@ -1,0 +1,5 @@
+export {
+  CMS_MENU_ADD,
+  CMS_MENU_EDIT,
+  CMS_MENU_REMOVE,
+} from './menuMutations';
