+++ frontend/plugins/content_ui/src/modules/cms/menus/graphql/queries/index.ts
@@ -1,0 +1,9 @@
+export { CMS_MENU_LIST } from './menuQueries';
+export {
+  MENU_CATEGORIES_QUERY,
+  MENU_CUSTOM_TYPES_QUERY,
+  MENU_DETECT_QUERY,
+  MENU_PAGES_QUERY,
+  MENU_POSTS_QUERY,
+  MENU_TAGS_QUERY,
+} from './menuDrawerQueries';
