+++ frontend/plugins/content_ui/src/modules/cms/menus/graphql/queries/menuDrawerQueries.ts
@@ -1,7 +1,7 @@
 import { gql } from '@apollo/client';
 
 export const MENU_PAGES_QUERY = gql`
-  query MenuPages($clientPortalId: String!, $limit: Int) {
+  query cmsMenuPages($clientPortalId: String!, $limit: Int) {
     cmsPageList(clientPortalId: $clientPortalId, limit: $limit) {
       pages {
         _id
@@ -13,7 +13,11 @@ export const MENU_PAGES_QUERY = gql`
 `;
 
 export const MENU_POSTS_QUERY = gql`
-  query MenuPosts($clientPortalId: String!, $limit: Int, $type: String) {
+  query cmsMenuPosts(
+    $clientPortalId: String!
+    $limit: Int
+    $type: String
+  ) {
     cmsPostList(clientPortalId: $clientPortalId, limit: $limit, type: $type) {
       posts {
         _id
@@ -25,7 +29,7 @@ export const MENU_POSTS_QUERY = gql`
 `;
 
 export const MENU_CUSTOM_TYPES_QUERY = gql`
-  query MenuCustomTypes($clientPortalId: String) {
+  query cmsMenuCustomTypes($clientPortalId: String) {
     cmsCustomPostTypes(clientPortalId: $clientPortalId) {
       _id
       code
@@ -35,7 +39,7 @@ export const MENU_CUSTOM_TYPES_QUERY = gql`
 `;
 
 export const MENU_CATEGORIES_QUERY = gql`
-  query MenuCategories($clientPortalId: String!, $limit: Int) {
+  query cmsMenuCategories($clientPortalId: String!, $limit: Int) {
     cmsCategories(clientPortalId: $clientPortalId, limit: $limit) {
       list {
         _id
@@ -47,7 +51,7 @@ export const MENU_CATEGORIES_QUERY = gql`
 `;
 
 export const MENU_TAGS_QUERY = gql`
-  query MenuTags($clientPortalId: String, $language: String) {
+  query cmsMenuTags($clientPortalId: String, $language: String) {
     cmsTags(clientPortalId: $clientPortalId, language: $language) {
       tags {
         _id
@@ -59,7 +63,7 @@ export const MENU_TAGS_QUERY = gql`
 `;
 
 export const MENU_DETECT_QUERY = gql`
-  query MenuDetect($clientPortalId: String!, $limit: Int) {
+  query cmsMenuDetect($clientPortalId: String!, $limit: Int) {
     cmsPageList(clientPortalId: $clientPortalId, limit: $limit) {
       pages {
         _id
